requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
django-environ==0.11.2
gunicorn==23.0.0
idna==3.10
numpy==2.1.2
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
